requirements.txt
astroid==2.1.0
autopep8==1.4.3  
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.39.1
django-appconf==1.0.4
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-imagekit==4.0.2
django-starcross-gallery==1.0.8
django-tinymce4-lite==1.7.4
dj-database-url==0.5.0
gunicorn==19.9.0
idna==2.8
isort==4.3.4
jsmin==2.2.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
pilkit==2.0
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.9
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.1
typed-ast==1.1.1
urllib3==1.24.2
whitenoise==5.1.0
wrapt==1.11.0
